+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [1.1.0] - 2022-04-16
 
+### Added
+
+* If elements are removed from the DOM, they are also forgotten by Scroll Appear so they can be garbage collected
+
 ### Changed
 
 * Items' `delay` and `queue` properties are now read when they are used, instead of set permanently when the item is initialised